Move schemas to separate package
