+++ backend/schemas/__init__.py
@@ -1,0 +1,3 @@
+from .schema import *
+from .entity import *
+from .traceability import *
